+++ Samples/iOS-ObjectiveCpp-NoModules/App/Sources/AppDelegate.h
@@ -2,6 +2,4 @@
 
 @interface AppDelegate : UIResponder <UIApplicationDelegate>
 
-@property (nonatomic, strong) UIWindow *window;
-
 @end